Rename CommentsContainer to CommentThreadContainer
