+++ react-app/src/components/CommentThreadContainer/index.js
@@ -1,0 +1,3 @@
+import CommentThreadContainer from './CommentThreadContainer';
+
+export default CommentThreadContainer;
