--- react-app/src/components/CommentsContainer/index.js
@@ -1,3 +1,0 @@
-import CommentsContainer from './CommentsContainer';
-
-export default CommentsContainer;
