+++ react-app/src/components/PostPage/PostPage.js
@@ -4,7 +4,7 @@ import { useParams } from 'react-router-dom';
 
 import { getPostById } from '../../store/posts';
 import Post from '../Post';
-import CommentsContainer from '../CommentsContainer';
+import CommentThreadContainer from '../CommentThreadContainer';
 
 const PostPage = () => {
   const { postId } = useParams();
@@ -32,7 +32,7 @@ const PostPage = () => {
   return (
     <>
       <Post post={post} userId={user ? user.id : null} />
-      <CommentsContainer threads={post.threads} />
+      <CommentThreadContainer threads={post.threads} />
     </>
   );
 };
